+++ src/kwja/datamodule/datasets/word_inference_dataset.py
@@ -18,6 +18,7 @@ import kwja
 from kwja.datamodule.datasets.base_dataset import BaseDataset
 from kwja.datamodule.examples import CohesionTask
 from kwja.datamodule.extractors import BridgingExtractor, CoreferenceExtractor, PasExtractor
+from kwja.utils.progress_bar import track
 from kwja.utils.sub_document import extract_target_sentences
 
 logger = logging.getLogger(__name__)
@@ -51,10 +52,12 @@ class WordInferenceDataset(BaseDataset):
     ) -> None:
         if knp_file is not None:
             with open(knp_file) as f:
-                documents = [Document.from_knp(c) for c in chunk_by_document(f)]
+                documents = [Document.from_knp(c) for c in track(chunk_by_document(f), description="Loading documents")]
         elif juman_file is not None:
             with open(juman_file) as f:
-                documents = [Document.from_jumanpp(c) for c in chunk_by_document(f)]
+                documents = [
+                    Document.from_jumanpp(c) for c in track(chunk_by_document(f), description="Loading documents")
+                ]
         else:
             documents = self._create_documents_from_texts(list(texts), doc_id_prefix)
         super().__init__(documents, document_split_stride, model_name_or_path, max_seq_length, tokenizer_kwargs or {})
@@ -109,7 +112,7 @@ class WordInferenceDataset(BaseDataset):
         if doc_id_prefix is None:
             doc_id_prefix = datetime.now().strftime("%Y%m%d%H%M")
         doc_id, sid = f"{doc_id_prefix}-0", f"{doc_id_prefix}-0-0"
-        for text in texts:
+        for text in track(texts, description="Loading documents"):
             if text.startswith("#"):
                 sentence = Sentence.from_raw_text(text)
                 doc_id, sid = sentence.doc_id, sentence.sid
@@ -145,7 +148,7 @@ class WordInferenceDataset(BaseDataset):
     def _load_examples(self, documents: List[Document]) -> List[WordInferenceExample]:
         examples = []
         idx = 0
-        for document in documents:
+        for document in track(documents, description="Loading examples"):
             encoding: Encoding = self.tokenizer(
                 " ".join(m.text for m in document.morphemes),
                 padding=PaddingStrategy.MAX_LENGTH,